Add a stripHeaders method.
Sometimes, you just want to be able to strip headers/comments from an
image without bothering with converting the image format or mucking with
file permissions, etc.  By creating a seperate method for stripping
headers, it also provides us with the ability to replace an
ImagickMagick lossy stripper with one that removes the headers without
re-encoding the jpeg or dropping color profiles.

A test image was added for testing exif header removal.  Image Magick
and the Imagick extension have very poor support for image properties
such as EXIF.  I have not found a way to get EXIF headers to persist on
an image after writeImage() is called.  So in order to test this, we
need to start with the exif headers on the image.

At some point I would definitely like to see us replace the EXIF tool,
but the existing options are poor:
* https://packagist.org/packages/phpexiftool/phpexiftool
  * Uses the CLI to access a perl tool to do the actual work
  * Has strong dependencies on supplying a Logger that we absolutely
    don't need.
* https://packagist.org/packages/lsolesen/pel
  * Old package that hasn't been updated recently
  * Horrible class structure that has not been updated to work well with
    composer
  * Strange localization effort wrapped around it
* https://packagist.org/packages/miljar/php-exif
  * Only read support, but he has an issue in for write support.  This
    package could be good if he adds write support and removes the
    dependency on the PECL exif extension.